# src/main/java/com/github/bondarevv23/forum/controller/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.github.bondarevv23.forum.controller.api;

import com.github.bondarevv23.forum.domain.generated.ErrorResponse;
import com.github.bondarevv23.forum.domain.generated.NotFoundResponse;
import com.github.bondarevv23.forum.domain.generated.UserDTO;
import com.github.bondarevv23.forum.domain.generated.UserRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.4.0")
@Validated
@Controller
@Tag(name = "User", description = "User service endpoints")
public interface UserApi {

    /**
     * DELETE /api/v1/users/{id} : delete user
     * delete user by id
     *
     * @param id id to delete (required)
     * @return user successfully deleted (status code 200)
     *         or error converting the passed parameter to id (status code 400)
     *         or unknown user id (status code 404)
     */
    @Operation(
        operationId = "deleteUserById",
        summary = "delete user",
        description = "delete user by id",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "user successfully deleted"),
            @ApiResponse(responseCode = "400", description = "error converting the passed parameter to id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "unknown user id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteUserById(
        @Parameter(name = "id", description = "id to delete", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /api/v1/users/ : get all users
     * get all stored users
     *
     * @return list of all users (status code 200)
     */
    @Operation(
        operationId = "findAllUsers",
        summary = "get all users",
        description = "get all stored users",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "list of all users", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/users/",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<UserDTO>> findAllUsers(
        
    );


    /**
     * GET /api/v1/users/{id} : get user
     * get user by id
     *
     * @param id id to search (required)
     * @return successfull request (status code 200)
     *         or user with passed id doesn&#39;t exist (status code 404)
     *         or error converting the passed parameter to id (status code 400)
     */
    @Operation(
        operationId = "findUserById",
        summary = "get user",
        description = "get user by id",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "successfull request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "user with passed id doesn't exist", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "error converting the passed parameter to id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserDTO> findUserById(
        @Parameter(name = "id", description = "id to search", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * POST /api/v1/users/ : rigister new user
     * register new user by passed data
     *
     * @param userRequest new user data (required)
     * @return user successfully created (status code 200)
     *         or invalid data passed (status code 400)
     *         or the request conflicts with the server state (status code 409)
     */
    @Operation(
        operationId = "register",
        summary = "rigister new user",
        description = "register new user by passed data",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "user successfully created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "invalid data passed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "the request conflicts with the server state", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/users/",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<UserDTO> register(
        @Parameter(name = "UserRequest", description = "new user data", required = true) @Valid @RequestBody UserRequest userRequest
    );


    /**
     * PUT /api/v1/users/{id} : update user
     * update user by id
     *
     * @param id id to update (required)
     * @param userRequest updated user data (required)
     * @return user successfully updated (status code 200)
     *         or invalid data passed (status code 400)
     *         or user with passed id doesn&#39;t exist (status code 404)
     *         or the request conflicts with the server state (status code 409)
     */
    @Operation(
        operationId = "updateUserById",
        summary = "update user",
        description = "update user by id",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "user successfully updated"),
            @ApiResponse(responseCode = "400", description = "invalid data passed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "user with passed id doesn't exist", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "the request conflicts with the server state", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/users/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> updateUserById(
        @Parameter(name = "id", description = "id to update", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "UserRequest", description = "updated user data", required = true) @Valid @RequestBody UserRequest userRequest
    );

}